+++ react/components/SearchFilterBar.js
@@ -1,6 +1,10 @@
 import React from 'react'
 import { defineMessages, injectIntl } from 'react-intl'
 import { Input } from 'vtex.styleguide'
+import classNames from 'classnames'
+import { useCssHandles } from 'vtex.css-handles'
+
+const CSS_HANDLES = ['filterSearch']
 
 const messages = defineMessages({
   searchPlaceHolder: {
@@ -10,13 +14,17 @@ const messages = defineMessages({
 })
 
 const SearchFilterBar = ({ title, handleChange, intl }) => {
+  const handles = useCssHandles(CSS_HANDLES)
+
   return (
-    <Input
-      onChange={e => handleChange(e.target.value.toLowerCase())}
-      placeholder={intl.formatMessage(messages.searchPlaceHolder, {
-        filterName: title,
-      })}
-    />
+    <div className={classNames('mb3', handles.filterSearch)}>
+      <Input
+        onChange={e => handleChange(e.target.value.toLowerCase())}
+        placeholder={intl.formatMessage(messages.searchPlaceHolder, {
+          filterName: title,
+        })}
+      />
+    </div>
   )
 }