+++ CHANGELOG.md
@@ -8,8 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ### Added
-- Search input for facets.
-- Prop `orderFacetsBy` to `SearchResultFlexible`.
+- Facets are now searchable.
+- Facets ordering.
 
 ### Fixed
 - `showFacetQuantity` was not working on mobile.