+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Changed
-- "Filter by" message to "Filtering by".
+- "Filter by" message to "Filtered by".
 
 ## [3.46.0] - 2020-02-04
 ### Added