+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [3.47.0] - 2020-02-07
+
 ## [3.46.1] - 2020-02-06
 ### Changed
 - "Filter by" message to "Filtered by".